Server/gRPC/Python/requirements.txt
grpcio-tools
protobuf >= 3.13.0
